+++ server.js
@@ -3,16 +3,16 @@ const path = require("path");
 // const { PORT } = require("./config/keys");
 const app = express();
 
-require("./startUp/db")();
-require("./startUp/parser")(app);
-require("./startUp/routes")(app);
+require("./src/startUp/db")();
+require("./src/startUp/parser")(app);
+require("./src/startUp/routes")(app);
 
-if (process.env.NODE_ENV === "production") {
-  app.use(express.static("client/build"));
-  app.get("*", (req, res) => {
-    res.sendFile(path.resolve(__dirname, "client", "build", "index.html"));
-  });
-}
+// if (process.env.NODE_ENV === "production") {
+//   app.use(express.static("client/build"));
+//   app.get("*", (req, res) => {
+//     res.sendFile(path.resolve(__dirname, "client", "build", "index.html"));
+//   });
+// }
 
 const PORT = process.env.PORT || 5000;
