+++ src/modules/customers/customer.model.js
@@ -1,23 +1,23 @@
 const mongoose = require("mongoose");
 const Schema = mongoose.Schema;
-const Joi = require('joi');
+const Joi = require("joi");
 
 const customerSchema = new Schema({
   user: {
     type: Schema.Types.ObjectId,
-    ref: "user"
+    ref: "user",
   },
   name: {
     type: String,
     minlength: 1,
     maxlength: 50,
-    required: true
+    required: true,
   },
   phone: {
     type: String,
     required: true,
     minlength: 5,
-    maxlength: 50
+    maxlength: 50,
   },
   address: {
     type: [
@@ -26,50 +26,37 @@ const customerSchema = new Schema({
           type: String,
           minlength: 5,
           maxlength: 255,
-          required: true
+          required: true,
         },
         number: {
           type: Number,
           min: 0,
-          required: true
+          required: true,
         },
         location: {
           type: String,
           minlength: 5,
           maxlength: 255,
-          required: true
-        }
-      }
+          required: true,
+        },
+      },
     ],
-  }
+  },
 });
 const validateCustomer = {
   customer: Joi.object().keys({
     _id: Joi.objectId(),
-    name: Joi.string()
-      .min(1)
-      .max(50)
-      .required(),
-    phone: Joi.string()
-      .min(5)
-      .max(50)
-      .required()
+    name: Joi.string().min(1).max(50).required(),
+    phone: Joi.string().min(5).max(50).required(),
   }),
   address: Joi.object().keys({
     _id: Joi.objectId(),
-    street: Joi.string()
-      .min(5)
-      .max(255)
-      .required(),
-    location: Joi.string()
-      .min(5)
-      .max(255)
-      .required(),
+    street: Joi.string().min(5).max(255).required(),
+    location: Joi.string().min(5).max(255).required(),
     number: Joi.number().min(0).required(),
-  })
-}
+  }),
+};
 const Customer = new mongoose.model("customer", customerSchema);
 
 exports.validateCustomer = validateCustomer;
 exports.Customer = Customer;
-
