+++ src/startUp/parser.js
